+++ scheduler.py
@@ -14,8 +14,6 @@ class Scheduler:
     count=0
     def __init__(self):
         self.schedules = []
-        self.apps = {}
-        self.app_stack = []
 
     def start(self):
         self.start = time.ticks_ms()
@@ -34,25 +32,3 @@ class Scheduler:
                     schedule.callback(t)
                     schedule.lastrun = tm
 
-    def add_app(self, app):
-        self.apps[app.name] = app
-
-    def enable_app(self, name):
-        for n in self.app_stack:
-            app = self.apps[n]
-            app.disable()
-        app = self.apps[name]
-        self.app_stack.append(name)
-        app.enable()
-
-    def disable_app(self, name):
-        for n in self.app_stack:
-            if n == name:
-                app = self.apps[n]
-                app.disable()
-        self.app_stack.remove(name)
-
-        if len(self.app_stack)>0:
-            n = self.app_stack[-1]
-            app = self.apps[n]
-            app.enable()