+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Change Log
 
-2021/09/19: Added 'Pomodoro timer' app
-2021/09/19: Updated scheduler to know about "apps"
+* **2021/10/01**: Improved 'app' functionality
+  Top button switches between apps
+* **2021/09/19**: Added 'Pomodoro timer' app
+* **2021/09/19**: Updated scheduler to know about "apps"